requirements.txt
# Core Streamlit and UI
streamlit>=1.28.0

# LangChain and MistralAI Integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-mistralai>=0.0.1

# AssemblyAI for Transcription
assemblyai>=0.21.0

# YouTube Processing
yt-dlp>=2023.12.30

# HTTP and API Requests
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# File and Audio Processing
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Data Handling and Utilities
pandas>=2.0.0
numpy>=1.24.0

# Progress and Status Indicators
tqdm>=4.66.0

# Error Handling and Logging
loguru>=0.7.0

# Optional: Enhanced UI Components
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0
